+++ NEWS.rst
@@ -3,6 +3,18 @@
 Unreleased
 =============================
 
+Removals
+------------------------------
+* `(defn/a …)` is now `(defn :async …)`.
+* `(fn/a …)` is now `(fn :async …)`.
+* `(with/a […] …)` is now `(with [:async …] …)`.
+
+   * As with `for`, `:async` must precede each name to be bound
+     asynchronously, because you can mix synchronous and asynchronous
+     types.
+
+* `(yield-from …)` is now `(yield :from …)`.
+
 New Features
 ------------------------------
 * You can now set `repl-ps1` and `repl-ps2` in your `HYSTARTUP` to customize